+++ paicli-server/src/test/java/com/paicli/platform/server/security/WebSecurityIntegrationTest.java
@@ -99,7 +99,7 @@ class WebSecurityIntegrationTest {
                 .andExpect(content().string(org.hamcrest.Matchers.containsString(
                         "data-effort=\"low\"")))
                 .andExpect(content().string(org.hamcrest.Matchers.containsString(
-                        "20260804-workbench-batch-delete-1")))
+                        "20260808-prd-analysis-1")))
                 .andExpect(content().string(org.hamcrest.Matchers.containsString(
                         "id=\"scheduleForm\"")))
                 .andExpect(content().string(org.hamcrest.Matchers.containsString(